BuiltinPlugins.IServerReporter is now BuildInServices.IServerReporterService
This removes the server reporter as a "built in plugin" as this was misleading. We don't want to be using the plugin terminology anymore other than for game servers, and the reporter was actually just a contained service.

I've also added preparation work for a fast-DL mechanism, and reflected the code in the reporter and webservice accordingly.